Replace some methods i removed with the new query methods
