+++ src/sentry/services/hybrid_cloud/user/__init__.py
@@ -2,17 +2,16 @@ from __future__ import annotations
 
 import datetime
 from abc import abstractmethod
-from dataclasses import dataclass, fields
+from dataclasses import dataclass
 from enum import IntEnum
-from typing import TYPE_CHECKING, Any, FrozenSet, Iterable, List, Optional, TypedDict
+from typing import TYPE_CHECKING, FrozenSet, List, Optional, TypedDict
 
-from sentry.db.models import BaseQuerySet
 from sentry.services.hybrid_cloud import InterfaceWithLifecycle, silo_mode_delegation, stubbed
 from sentry.services.hybrid_cloud.filter_query import FilterQueryInterface
 from sentry.silo import SiloMode
 
 if TYPE_CHECKING:
-    from sentry.models import Group, User
+    from sentry.models import Group
 
 
 @dataclass(frozen=True, eq=True)
@@ -154,64 +153,6 @@ class UserService(
             return users[0]
         else:
             return None
-
-    @classmethod
-    def serialize_user(cls, user: User) -> APIUser:
-        args = {
-            field.name: getattr(user, field.name)
-            for field in fields(APIUser)
-            if hasattr(user, field.name)
-        }
-        args["pk"] = user.pk
-        args["display_name"] = user.get_display_name()
-        args["label"] = user.get_label()
-        args["is_superuser"] = user.is_superuser
-        args["is_sentry_app"] = user.is_sentry_app
-        args["password_usable"] = user.has_usable_password()
-        args["emails"] = frozenset([email.email for email in user.get_verified_emails()])
-
-        # And process the _base_user_query special data additions
-        permissions: FrozenSet[str] = frozenset({})
-        if hasattr(user, "permissions") and user.permissions is not None:
-            permissions = frozenset(user.permissions)
-        args["permissions"] = permissions
-
-        roles: FrozenSet[str] = frozenset({})
-        if hasattr(user, "roles") and user.roles is not None:
-            roles = frozenset(flatten(user.roles))
-        args["roles"] = roles
-
-        useremails: FrozenSet[APIUserEmail] = frozenset({})
-        if hasattr(user, "useremails") and user.useremails is not None:
-            useremails = frozenset(
-                {
-                    APIUserEmail(
-                        id=e["id"],
-                        email=e["email"],
-                        is_verified=e["is_verified"],
-                    )
-                    for e in user.useremails
-                }
-            )
-        args["useremails"] = useremails
-        avatar = user.avatar.first()
-        if avatar is not None:
-            avatar = APIAvatar(
-                id=avatar.id,
-                file_id=avatar.file_id,
-                ident=avatar.ident,
-                avatar_type=avatar.avatar_type,
-            )
-        args["avatar"] = avatar
-        return APIUser(**args)
-
-
-def flatten(iter: Iterable[Any]) -> List[Any]:
-    return (
-        ((flatten(iter[0]) + flatten(iter[1:])) if len(iter) > 0 else [])
-        if type(iter) is list or isinstance(iter, BaseQuerySet)
-        else [iter]
-    )
 
 
 def impl_with_db() -> UserService:
